Remove the authorization configuration and mssts webviewclient as the webviewclient layer does not need to know the difference between V1 and V2.
