+++ common/src/main/java/com/microsoft/identity/common/internal/ui/embeddedwebview/challengehandlers/PKeyAuthChallengeHandler.java
@@ -36,7 +36,7 @@ final class PKeyAuthChallengeHandler {
         mResponse = response;
     }
 
-    void handleChallenge() {
+    void processPkeyAuthChallenge() {
         //TODO
     }
 }
